+++ BE/error/src/main/java/com/example/demo/schedule/infrastructure/persistence/ScheduleEntity.java
@@ -1,4 +1,4 @@
-package com.example.demo.schedule.persistence;
+package com.example.demo.schedule.infrastructure;
 
 import com.example.demo.auth.persistence.MemberEntity;
 import com.example.demo.common.persistence.BaseEntity;
